% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex.R
\name{flex}
\alias{flex}
\title{Select one or more colors from the Flexoki palette.}
\usage{
flex(col, sat, code = "hex", palette = flexoki)
}
\arguments{
\item{col}{A character string specifying one or more hues of colors. See \code{Details} for options.}

\item{sat}{A number specifying the saturation. \code{NA} is allowed when \code{col = "paper"} or \code{col = "base"}.}

\item{code}{A character string specifying the form of color code (default = \code{"hex"}).}

\item{palette}{Name of the chosen color palette; \code{flexoki} is the only option.}
}
\value{
A vector with one or more color codes.
}
\description{
Select one or more colors from the Flexoki palette.
}
\details{
Additional details...
}
\examples{
flex(c("blue", "red"), c(100))
flex(c("blue"), c(100, 200))
flex(c("blue", "red", "green"), c(100, 200, 100))
}
