% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexoki-package.R
\docType{package}
\name{flexoki-package}
\alias{flexoki}
\alias{flexoki-package}
\title{flexoki: Creates Color Palettes from the Flexoki Theme}
\description{
This package extands the "Flexoki" color palette to the R computing environment.
}
\author{
\strong{Maintainer}: Mark Scheuerell \email{scheuerl@uw.edu} (\href{https://orcid.org/0000-0002-8284-1254}{ORCID})

}
\keyword{internal}
