% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{flexoki}
\alias{flexoki}
\title{'Flexoki' Color Palettes}
\format{
list of \link[palette:palette]{palette::palette} vectors
}
\usage{
flexoki
}
\description{
Color palettes from Steph Ango's \url{https://stephango.com/flexoki}.
}
\examples{
plot(flexoki$dark)
plot(flexoki$light)

}
\concept{colors}
\keyword{datasets}
