% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_flexoki_c}
\alias{scale_color_flexoki_c}
\alias{scale_fill_flexoki_c}
\alias{scale_colour_flexoki_c}
\title{Continuous 'Flexoki' Color Scales for \code{ggplot2}}
\usage{
scale_color_flexoki_c(palette = "red", ..., reverse = FALSE)

scale_fill_flexoki_c(palette = "red", ..., reverse = FALSE)

scale_colour_flexoki_c(palette = "red", ..., reverse = FALSE)
}
\arguments{
\item{palette}{palette from \code{names(flexoki)} to use}

\item{...}{arguments passed on to \code{ggplot2::discrete_scale()}}

\item{reverse}{Should the vector be reversed? Default is \code{FALSE}.}
}
\value{
A \link[ggplot2:ggplot2-ggproto]{ggplot2::Scale}
}
\description{
Continuous 'Flexoki' Color Scales for \code{ggplot2}
}
\examples{
ggplot2::faithfuld |>
  ggplot2::ggplot() +
  ggplot2::geom_tile(ggplot2::aes(waiting, eruptions, fill = density)) +
  scale_fill_flexoki_c(palette = 'red')

}
