% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_flexoki_dark}
\alias{theme_flexoki_dark}
\title{Dark Theme}
\usage{
theme_flexoki_dark(base_size = 12, base_family = "mono", ...)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{...}{additional parameters}
}
\value{
ggplot2 theme
}
\description{
Dark Theme
}
\examples{
ggplot2::mpg |>
  ggplot2::ggplot() +
  ggplot2::geom_point(ggplot2::aes(displ, hwy, colour = class)) +
  scale_color_flexoki_d(palette = 'dark') +
  theme_flexoki_dark()

}
\concept{theme}
