% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex.R
\name{flex}
\alias{flex}
\title{Select one or more colors from the Flexoki palette.}
\usage{
flex(col, sat = 400, code = "hex", palette = flexoki)
}
\arguments{
\item{col}{Character string specifying one or more hues of colors (see Details).}

\item{sat}{Number specifying the saturation level. If \code{col = "paper"} or \code{col = "black"} this defaults to \code{sat = NA} (see Details).}

\item{code}{Character string specifying the form of color code (default = \code{"hex"}).}

\item{palette}{Name of the chosen color palette; \code{flexoki} is the only option at this time.}
}
\value{
For \code{code = "hex"}, a scalar (or vector) of one (or more) hexadecimal color code(s).

For \code{code = "rgb"}, a tibble with columns \code{c("red", "green", "blue")}
}
\description{
Select one or more colors from the Flexoki palette.
}
\details{
There are three different options for monochromatic base hues: off-white (\code{col = "paper"}), black (\code{col = "black"}), and a 12-level gray-scale (\code{col = "gray"} or \code{col = "base"}).

There are several options for accent colors, each specified using the \code{col} argument and offered in 12 increasing levels of saturation. The color options are:

\code{c("red", "orange", "yellow", "green", "cyan", "blue", "purple", "magenta")}

The level of saturation is denoted by an integer from the following set, ordered from lightest to darkest:

\code{c(50, 100, 150, 200, 300, 400, 500, 600, 700, 800, 850, 900, 950)}.

The function will work with vector arguments for \code{col} and \code{sat}. If the vectors are the same length, the function assumes a 1:1 correspondence between a specified color and the saturation, such that \code{col = c("red", "blue")} and \code{sat = c(100, 200)} would yield the hexadecimal equivalent of \code{"red-100"} (\verb{#FFCABB}) and \code{"blue-200"} (\verb{#92BFDB}). If one of the arguments is a vector and the other is a scalar, the shorter argument is recycled (see examples).
}
\examples{
## two different colors with same saturation levels
#> flex(c("blue", "red"), 100)

## two different colors with same saturation levels in RGB
flex(c("blue", "red"), 100, code = "rgb")

## gray with different saturation levels
flex("gray", c(100, 200))

## three colors with specific saturation levels
flex(c("blue", "red", "green"), c(100, 200, 300))

}
