% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexoki-package.R
\docType{package}
\name{flexoki-package}
\alias{flexoki-package}
\title{flexoki: Inky Color Schemes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides color palettes designed to be reminiscent of text on paper. The color schemes were taken from \url{https://stephango.com/flexoki}. Includes discrete, continuous, and binned scales that are not necessarily color-blind friendly. Simple scale and theme functions are available for use with 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/christopherkenny/flexoki}
  \item \url{https://christophertkenny.com/flexoki/}
  \item Report bugs at \url{https://github.com/christopherkenny/flexoki/issues}
}

}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{christopherkenny@fas.harvard.edu} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

}
\keyword{internal}
