% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexoki-package.R
\docType{package}
\name{flexoki-package}
\alias{flexoki}
\alias{flexoki-package}
\title{flexoki: Creates Color Palettes from the Flexoki Theme}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package extands the "Flexoki" color palette to the R computing environment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdscheuerell/flexoki}
  \item Report bugs at \url{https://github.com/mdscheuerell/flexoki/issues}
}

}
\author{
\strong{Maintainer}: Mark Scheuerell \email{scheuerl@uw.edu} (\href{https://orcid.org/0000-0002-8284-1254}{ORCID})

}
\keyword{internal}
